+++ lib/vagrant/devcommands/runner/chain.rb
@@ -3,6 +3,8 @@ module VagrantPlugins
     module Runner
       # Chain runner
       class Chain
+        UTIL = VagrantPlugins::DevCommands::Util
+
         def initialize(plugin, argv, env, registry)
           @plugin   = plugin
           @argv     = argv
@@ -29,10 +31,22 @@ module VagrantPlugins
         private
 
         def argv_for(command_def)
-          return @argv unless command_def.key?(:argv)
-          return @argv unless command_def[:argv].is_a?(Array)
+          argv = patch_machine(@argv.dup, command_def)
 
-          @argv + command_def[:argv]
+          return argv unless command_def.key?(:argv)
+          return argv unless command_def[:argv].is_a?(Array)
+
+          argv + command_def[:argv]
+        end
+
+        def patch_machine(argv, command_def)
+          return argv unless command_def.key?(:machine)
+
+          if UTIL.machine_name?(argv[0].to_s, @env.machine_index)
+            argv
+          else
+            [command_def[:machine].to_s] + argv
+          end
         end
 
         def runnable_for(command_def)