+++ test/acceptance/test.sh
@@ -75,9 +75,30 @@ expect_2='[secondary.vagrant.devcommands initial initial]'
 expect_3='[secondary.vagrant.devcommands param param]'
 
 [[ "${result}" == *"${expect_1}"*"${expect_2}"*"${expect_3}"* ]] || {
-  echo 'Machine passed using argv not taking precedence over configuration...'
+  echo 'Machine passed using argv not taking precedence command over configuration...'
   echo "Got result: '${result}'"
   exit 1
 }
 
+result=$(trim "$(bundle exec vagrant run chainmachine)")
+expect_1='primary.vagrant.devcommands'
+expect_2='secondary.vagrant.devcommands'
+
+[[ "${result}" == *"${expect_1}"*"${expect_2}"* ]] || {
+  echo 'Chain machine configuration did not take precedence over command configuration...'
+  echo "Got result: '${result}'"
+  exit 1
+}
+
+result=$(trim "$(bundle exec vagrant run secondary chainmachine)")
+expect_1='secondary.vagrant.devcommands'
+expect_2='secondary.vagrant.devcommands'
+
+[[ "${result}" == *"${expect_1}"*"${expect_2}"* ]] || {
+  echo 'Machine passed using argv not taking precedence chain over configuration...'
+  echo "Got result: '${result}'"
+  exit 1
+}
+
+# finish
 echo 'Success!'